+++ components/Header.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from "react";
 import Image from "next/image";
 import styles from "../styles/LoginDropdown.module.css";
-import router from "next/router";
+import DropdownCard from "./LoginDropdown";
 
 function Header({ isUserLoggedIn, navbarBackground }) {
   // localStorage.setItem('redirectUrl', window.location.href);
@@ -9,7 +9,6 @@ function Header({ isUserLoggedIn, navbarBackground }) {
   const [isMentorLoggedIn, setMentorLoggedIn] = useState(false);
   const [scrollY, setScrollY] = useState(0);
   const [navbarAppear, setNavbarAppear] = useState(false);
-  const [loginOption, setLoginOption] = useState(false);
 
   useEffect(() => {
     const userName = localStorage.getItem("user_name");
@@ -33,33 +32,12 @@ function Header({ isUserLoggedIn, navbarBackground }) {
     };
   }, []);
 
-  const handleLoginClick = () => {
-    setLoginOption(true);
-    setTimeout(() => {
-      setLoginOption(false);
-    }, 5000);
-  };
-
   const menuToggle = () => {
     if (navbarAppear === true) {
       setNavbarAppear(false);
     } else {
       setNavbarAppear(true);
     }
-  };
-
-  const userlogout = () => {
-    localStorage.clear();
-    setLoggedIn(false);
-    router.push("/");
-    window.location.reload();
-  };
-
-  const mentorlogout = () => {
-    localStorage.clear();
-    setMentorLoggedIn(false);
-    router.push("/");
-    window.location.reload();
   };
 
   return (
@@ -114,90 +92,10 @@ function Header({ isUserLoggedIn, navbarBackground }) {
                             Contact
                           </a>
                         </li>
-
-                        {isLoggedIn || isUserLoggedIn || isMentorLoggedIn ? (
-                          <li>
-                            <div className={styles.loginOption}>
-                              <button
-                                onClick={handleLoginClick}
-                                className={styles.userName}
-                              >
-                                <img
-                                  style={{
-                                    width: "35px",
-                                    height: "auto",
-                                    borderRadius: "50%",
-                                  }}
-                                  src={
-                                    localStorage.getItem("user_picture") ||
-                                    localStorage.getItem("mentor_picture") ||
-                                    "assets/img/icon/no-profile-picture.png"
-                                  }
-                                  alt="not found"
-                                />
-                              </button>
-
-                              {loginOption && (
-                                <div className="login-optionslist">
-                                  <button
-                                    id="login-buttons"
-                                    style={{ marginTop: "20px" }}
-                                    onClick={() => {
-                                      if (isMentorLoggedIn) {
-                                        window.location.href = `/dashboard`;
-                                      } else {
-                                        window.location.href = `/`;
-                                      }
-                                    }}
-                                  >
-                                    Dashboard
-                                  </button>
-                                  <button
-                                    id="login-buttons"
-                                    onClick={() => {
-                                      if (isMentorLoggedIn) {
-                                        mentorlogout();
-                                      } else {
-                                        userlogout();
-                                      }
-                                    }}
-                                  >
-                                    Logout
-                                  </button>
-                                </div>
-                              )}
-                            </div>
-                          </li>
-                        ) : (
-                          <li>
-                            <div className={styles.loginOption}>
-                              <button
-                                className={styles.loginbutton}
-                                onClick={handleLoginClick}
-                              >
-                                Login
-                              </button>
-                              {loginOption && (
-                                <div className="login-optionslist">
-                                  <button
-                                    id="login-buttons"
-                                    onClick={handleLoginClick}
-                                  >
-                                    <a href="/userAuth" id="loginbtn">
-                                      User
-                                    </a>
-                                  </button>
-                                  <button
-                                    id="login-buttons"
-                                    onClick={handleLoginClick}
-                                  >
-                                    <a href="/mentorLogin">Mentor</a>
-                                  </button>
-                                </div>
-                              )}
-                            </div>
-                          </li>
-                        )}
+                        <DropdownCard
+                          isUserLoggedIn={isUserLoggedIn}
+                          isMentorLoggedIn={isMentorLoggedIn}
+                        />
                       </ul>
                     </nav>
                   </div>
