+++ styles/LoginDropdown.module.css
@@ -1,12 +1,12 @@
 /* LoginDropdown.module.css */
 
 .navLink:hover {
-  color: rgb(255, 255, 255) !important;
+  color: rgb(243, 243, 248) !important;
   transform: scale(1.05);
-  text-shadow: 0 0 10px rgba(255, 255, 255, 0.9),
+  /* text-shadow: 0 0 10px rgba(255, 255, 255, 0.9),
     0 0 20px rgba(255, 255, 255, 0.8), 0 0 30px rgba(255, 255, 255, 0.7),
     0 0 40px rgba(255, 255, 255, 0.6), 0 0 50px rgba(255, 255, 255, 0.5),
-    0 0 60px rgba(255, 255, 255, 0.4), 0 0 70px rgba(255, 255, 255, 0.3);
+    0 0 60px rgba(255, 255, 255, 0.4), 0 0 70px rgba(255, 255, 255, 0.3); */
 }
 
 @media (max-width: 768px) {
@@ -28,6 +28,7 @@
 .loginOption {
   display: flex;
   flex-direction: column;
+  width: auto;
   position: relative;
   gap: 1rem;
   border-radius: 4px;
@@ -48,14 +49,10 @@
   font-weight: 600;
 }
 
-.loginbutton {
-  width: fit-content;
-}
-
 .loginOption {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: 10px;
   transform: translateZ(-50px);
 }
 
@@ -64,15 +61,17 @@
 }
 
 .loginOption button {
-  background: none;
+  color: black;
   border: none;
   cursor: pointer;
-  border-radius: 5px;
+  border-radius: 7px;
+  padding: 5px;
   font-weight: 600;
 }
 
 .loginOption button:hover {
-  background-color: #6f42c1;
+  background-color: #9160ec;
+  color: aliceblue;
 }
 
 .loginOption .userPicture:hover {
