+++ styles/globals.css
@@ -18,6 +18,10 @@ a {
 * {
   box-sizing: border-box;
 }
+
+/* img {
+  display: inline;
+} */
 /*  */
 
 .slider-area .hero__caption.hero__caption2,
@@ -30,7 +34,7 @@ a {
   position: absolute;
   background-color: white;
   width: 100%;
-  height: 100vh;
+  height: 100%;
   top: 0;
   left: 0;
   z-index: 99;
@@ -250,10 +254,10 @@ a {
     display: flex;
   }
   .header-bottom .container-fluid .row div:nth-child(2) {
-    padding: 0;
+    /* padding: 0; */
     position: absolute;
     top: 90px;
-    width: 100% !important;
+    /* width: 100% !important; */
     max-width: 100% !important;
     left: 0;
   }
@@ -265,10 +269,10 @@ a {
     display: block !important;
   }
   .main-menu nav ul {
-    gap: 50px;
-    padding: 50px 35px;
+    gap: 40px;
+    padding: 60px 35px;
     width: 100%;
-    height: 380px;
+    height: 100%;
     background: linear-gradient(to bottom, #c054ff 0%, #5274ff 100%);
     display: flex;
     flex-direction: column;
@@ -283,6 +287,16 @@ a {
     margin-top: -120px;
   }
   .container {
+    padding: 20px;
+    max-width: 100% !important;
+    width: 100% !important;
+  }
+}
+
+@media (min-width: 992px) {
+  .container {
+    padding: 20px;
+    margin: 10px;
     padding: 50px;
     max-width: 100% !important;
     width: 100% !important;
@@ -1038,9 +1052,16 @@ textarea {
 }
 .login-optionslist {
   position: absolute;
-  top: 30px !important;
+  top: 40px !important;
   display: flex;
   flex-direction: column !important;
+  gap: 10px;
+  background-color: white;
+  padding: 7px;
+  margin: 5px;
+  border-radius: 10px;
+  right: 1px;
+  width: 100px;
 }
 
 /*Mentor CSS*/
@@ -1148,6 +1169,17 @@ textarea {
 }
 
 @media only screen and (max-width: 800px) {
+  .login-optionslist {
+    position: absolute;
+    top: 40px !important;
+    display: flex;
+    flex-direction: column !important;
+    gap: 5px;
+    background-color: white;
+    padding: 10px;
+    border-radius: 10px;
+  }
+
   .mentor-card {
     flex-direction: column;
     width: 100%;
@@ -1164,8 +1196,8 @@ textarea {
   }
 
   .mentor-card__photo {
-    width: 15vw;
-    height: 15vw;
+    width: 25vw;
+    height: 25vw;
     margin: auto;
   }
